todo_list_two

+++ myapp/views.py
@@ -1,35 +1,50 @@
-from django.shortcuts import render,redirect
+from rest_framework.decorators import api_view
+from rest_framework.response import Response
+from rest_framework import status
+
 from .models import Task
+from .serializers import TaskSerializer
 
-def task_list(request):
-    if request.method == "POST":
-        title=request.POST.get("title")
-        Task.objects.create(title=title)
-    
-        return redirect("/")
 
-    tasks = Task.objects.all()
-    return render(request, 'tasks.html', {'tasks': tasks})
+@api_view(['GET', 'POST'])
+def task_list_api(request):
 
-def complete_task():
-    task =Task.objects.get(id=pk)
-    task.complete=True
-    task.save()
-    return redirect("/")
+    if request.method == 'GET':
+        tasks = Task.objects.all()
+        serializer = TaskSerializer(tasks, many=True)
+        return Response(serializer.data)
 
-def delete_task(request,pk):
-    task=Task.objects.get(id=pk)
-    task.delete()
-    return redirect("/")
+    if request.method == 'POST':
+        serializer = TaskSerializer(data=request.data)
 
-def update_task(request,pk):
-    task=Task.objects.all(id=pk)
-    if request.methof== "POST":
-        title =request.POST.get("title")
-        task.title =title
-        title.save()
-        return redirect("/")
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
 
-    return render(request, "update.html", {"task": task})
-        
-    
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+@api_view(['GET', 'PUT', 'DELETE'])
+def task_detail_api(request, pk):
+
+    try:
+        task = Task.objects.get(pk=pk)
+    except Task.DoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+
+    if request.method == 'GET':
+        serializer = TaskSerializer(task)
+        return Response(serializer.data)
+
+    elif request.method == 'PUT':
+        serializer = TaskSerializer(task, data=request.data)
+
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    elif request.method == 'DELETE':
+        task.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
